Fixes: https://github.com/komuw/ong/issues/125

```sh
benchstat old.txt new.txt

name   old time/op    new time/op    delta
Enc-8    80.8ms ± 1%    82.5ms ± 2%   +2.04%  (p=0.000 n=19+18)

name   old alloc/op   new alloc/op   delta
Enc-8    33.6MB ± 0%    33.6MB ± 0%   +0.00%  (p=0.000 n=20+17)

name   old allocs/op  new allocs/op  delta
Enc-8      39.0 ± 0%      44.0 ± 0%  +12.82%  (p=0.000 n=20+20)
```