@komuw 2022-09-20
We won't do this. Too much work and we don't know if it is even working.
We'll instead wait for https://github.com/golang/go/issues/21865